back-end/requirements.txt
fastapi==0.111.0
llama_index==0.10.44
openai==1.30.5
pydantic==2.7.2
python-dotenv==1.0.1
setuptools==58.0.4
